Appease stickler and migrate time-domain test
