+++ tests/test_core.py
@@ -13,15 +13,6 @@ from pyam.core import _meta_idx, concat
 from pyam.utils import isstr
 from pyam.testing import assert_iamframe_equal
 
-from conftest import (
-    TEST_YEARS,
-    TEST_DTS,
-    TEST_TIME_STR,
-    TEST_TIME_STR_HR,
-    TEST_TIME_MIXED,
-    EXP_DATETIME_INDEX,
-)
-
 
 df_filter_by_meta_matching_idx = pd.DataFrame(
     [
@@ -312,27 +303,6 @@ def test_equals_raises(test_pd_df):
 def test_get_item(test_df, column):
     """Assert that getting a column from `data` via the direct getter works"""
     pdt.assert_series_equal(test_df[column], test_df.data[column])
-
-
-# TODO implement this parametrization as part of `conftest.py:test_df`
-@pytest.mark.parametrize(
-    "time, domain, index",
-    [
-        (TEST_YEARS, "year", pd.Int64Index([2005, 2010])),
-        (TEST_DTS, "datetime", pd.DatetimeIndex(TEST_DTS)),
-        (TEST_TIME_STR, "datetime", pd.DatetimeIndex(TEST_DTS)),
-        (TEST_TIME_STR_HR, "datetime", pd.DatetimeIndex(TEST_TIME_STR_HR)),
-        (TEST_TIME_MIXED, "mixed", pd.Index(TEST_TIME_MIXED)),
-    ],
-)
-def test_time_domain(test_pd_df, time, domain, index):
-    # assert that the time-domain and time-index attributes are set correctly
-    mapping = dict([(i, j) for i, j in zip(TEST_YEARS, time)])
-    df = IamDataFrame(data=test_pd_df.rename(mapping, axis="columns"))
-
-    assert df.time_col == "year" if domain == "year" else "time"
-    assert df.time_domain == domain
-    pdt.assert_index_equal(df.time, index)
 
 
 def test_index(test_df_year):
